+++ test/generators/query_generator_test.rb
@@ -30,6 +30,8 @@ class QueryGeneratorTest < Rails::Generators::TestCase
   end
 
   test "A new fixture file for the query is created" do
+    skip
+
     run_generator
 
     assert_file "test/fixtures/graphql/metaphysics/artist.yml" do |fixture|